requirements.txt
absl-py==0.9.0
altgraph==0.17
astor==0.8.1
Babel==2.9.1
backports.entry-points-selectable==1.1.0
bce-python-sdk==0.8.62
cachetools==4.2.2
certifi==2020.6.20
cfgv==3.3.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cssselect==1.1.0
cssutils==2.3.0
cycler==0.10.0
decorator==5.0.9
distlib==0.3.2
et-xmlfile==1.1.0
experimental==0.9.7
filelock==3.0.12
flake8==3.9.2
Flask==2.0.1
Flask-Babel==2.0.0
flatbuffers==2.0
future==0.18.2
gast==0.3.3
google-auth==1.35.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
identify==2.2.13
idna==3.2
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==1.6.0
infer==0.1
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lmdb==1.2.1
lxml==4.6.3
Markdown==3.2.2
MarkupSafe==2.0.1
matplotlib==3.3.2
mccabe==0.6.1
networkx==2.6.2
nltk==3.6.2
nodeenv==1.6.0
numpy==1.19.3
oauthlib==3.1.1
onnxruntime==1.8.1
opencv-python==4.4.0.46
opt-einsum==3.2.1
pandas==1.3.2
pefile==2021.5.24
pika==1.2.0
Pillow==8.0.1
platformdirs==2.2.0
pre-commit==2.14.0
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0
pycodestyle==2.7.0
pycryptodome==3.10.1
pyflakes==2.3.1
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytils==0.3
pytz==2021.1
PyWavelets==1.1.1
pywin32-ctypes==0.2.0
PyYAML==5.4.1
regex==2021.8.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.17.2
scipy==1.7.1
seaborn==0.11.2
Shapely==1.7.1
shellcheck-py==0.7.2.1
six==1.15.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
thop==0.0.31.post2005241907
tifffile==2021.8.8
toml==0.10.2
torch==1.9.0
torchvision==0.10.0
tqdm==4.46.0
typing-extensions==3.10.0.0
urllib3==1.26.6
virtualenv==20.7.2
visualdl==2.2.0
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.1.0
pip install paddlepaddle -i https://mirror.baidu.com/pypi/simple
